src/api/requirements.txt
ultralytics
fastapi
uvicorn
waitress
pillow
sqlalchemy